==> notebooks/01_extraction.ipynb <==
import json
import pandas as pd
from tqdm.notebook import tqdm
import re
import json

def show(x):
    print(x.shape)
    return x.head()

with open("../data/raw/science_parse_output.json") as file:
    lines = [json.loads(l) for l in tqdm(file.readlines())]
len(lines)

lines[0]

raw = pd.DataFrame(lines)\
.assign(title = lambda df: df.metadata.map(lambda x: x["title"]),
        abstract = lambda df: df.metadata.map(lambda x: x["abstractText"]),
        sections = lambda df: df.metadata.map(lambda x: x["sections"]),
        references = lambda df: df.metadata.map(lambda x: x["references"]),
        citations = lambda df: df.metadata.map(lambda x: x["referenceMentions"]),
)\

show(raw)

raw_na = raw.where(raw.applymap(lambda x: x is not None and x != []))\
.dropna()\
.assign(    
    title_clean = lambda df: df.title.str.replace(r"\W", "", regex = True).str.lower()
)\
.applymap(json.dumps)\
.groupby("title_clean").min()\
.applymap(json.loads)\
.reset_index()\
.reset_index()\
.rename(columns = {"index":"paper_id"})


show(raw_na)

papers = raw_na[["paper_id", "title", "abstract", "sections"]]\

show(papers)

raw_na.citations.map(len, na_action = "ignore").sum()

citations_raw = raw_na.citations.dropna().explode()
show(citations_raw)

citations = pd.DataFrame(citations_raw.tolist())\
.assign(paper_id = citations_raw.index)\
[["paper_id", "referenceID", "context", "startOffset", "endOffset"]]\
.rename(
    columns = {
        "referenceID":"internal_reference_id", 
        "startOffset":"start_offset", 
        "endOffset":"end_offset"
    }
)

show(citations)
citations.head(20)

citations.context[2]

references = raw_na.references.explode().reset_index()\
.assign(
    internal_reference_id = lambda df: df.groupby("index").cumcount(),
    title = lambda df: df.references.map(lambda x: x["title"], na_action = "ignore")
)\
.assign(
    title_clean = lambda df: df.title.str.replace(r"\W", "", regex = True).str.lower()
)\
.rename(columns = {"index":"paper_id"})


show(references)


# In some papers, there are references for which science-parse didn't parse all the 
# citations correctly. These papers have more references than citations.
# There are two papers (181 and 640) that apparently had some trouble parsing the references.

(citations.groupby(["paper_id", "internal_reference_id"])\
 .size().groupby("paper_id").size()
 - references.groupby(["paper_id", "internal_reference_id"])\
 .size().groupby("paper_id").size())\
.sort_values(ascending = False)

# The counts shown here are how many times a given paper appeared as reference in a paper, 
# parsed with science-parse

unique_references = references.groupby("title_clean")\
.agg(["min", "count"])\
.title\
.sort_values("count", ascending = False)\
.reset_index()\
.reset_index()\
.rename(columns = {"index":"reference_id", "min":"title", "count":"original_citations"})

show(unique_references)

# There are some rows for which there are discrepancies because science-parse failed to parse them correctly

full = citations.merge(
    references, 
    how = "left", 
    on = ["paper_id", "internal_reference_id"]
)\
.merge(unique_references.drop(columns = "title"), on = "title_clean", how = "left")

show(full)

full.loc[full.title_clean.isna()]

### just checking that everything makes sense

full.groupby(["reference_id", "paper_id"]).size()\
.sort_values(ascending = False)\
.rename("mentions")\
.reset_index()\
.groupby("reference_id").size()\
.sort_values(ascending = False)

full.groupby("reference_id").size().sort_values(ascending = False)

full.groupby(["title_clean", "paper_id"]).size()\
.sort_values(ascending = False)\
.rename("mentions")\
.reset_index()\
.groupby("title_clean").size()\
.sort_values(ascending = False)

full.groupby("title_clean").size().sort_values(ascending = False)

# final datasets

final_citations = full[
    [
        "paper_id", 
         "internal_reference_id", 
         "reference_id", 
         "context", 
         "start_offset", 
         "end_offset"
    ]
]\
.dropna()\
.assign(
    reference_id = lambda df: df.reference_id.map(
        int, 
        na_action = "ignore"
    )
)
show(final_citations)

final_papers = papers.assign(
    text = lambda df: df.sections.map(
        lambda x: " ".join([s["text"] for s in x]), 
        na_action = "ignore"
    )
)\
.assign(title = lambda df: df.title.str.replace(r"\s+", " ", regex = True).str.strip(),
        abstract = lambda df: df.abstract.str.replace(
            r"\s+", " ", 
            regex = True
        ).str.strip(),
        text = lambda df: df.text.str.replace(r"\s+", " ", regex = True).str.strip())
show(final_papers)

references_abstract = unique_references\
.merge(
    final_papers.assign(
        title_clean = lambda df: df.title.str.replace(r"\W", "", regex = True).str.lower()
    ),
    how = "left"
)\
[["reference_id", "title", "abstract", "original_citations", "title_clean"]]\
.drop_duplicates()

show(references_abstract)

# complete references from the arxiv dataset

titles = []
abstracts = []
with open("../data/raw/arxiv-metadata-oai-snapshot.json") as file:
    lines = file.readlines()
    for l in tqdm(lines):
        d = json.loads(l)
        titles.append(re.sub(r"\W", "", d["title"]).lower())
        abstracts.append(d["abstract"])

arxiv = pd.DataFrame(dict(title_clean = titles, abstract = abstracts))
show(arxiv)

full_references = references_abstract.merge(arxiv, on = "title_clean", how = "left")\
.assign(abstract = lambda df: df.abstract_y.combine_first(df.abstract_x))\
.drop(columns = ["abstract_x", "title_clean", "abstract_y"])\
.drop_duplicates()\
.assign(
    title = lambda df: df.title.str.replace(
        r"\s+", 
        " ", 
        regex = True
    ).str.strip(),
    abstract = lambda df: df.abstract.str.replace(
        r"\s+", " ", 
        regex = True
    ).str.strip()
)\
.groupby("reference_id").min()\
.reset_index()

show(full_references)
full_references.head(20)

# paper sections

paper_sections = final_papers\
.drop(columns = ["title", "text", "abstract"])\
.explode("sections")\
.assign(
    section_order = lambda df: df.groupby("paper_id").cumcount(),
    heading = lambda df: df.sections.map(lambda x: x["heading"], na_action = "ignore"),
    text = lambda df: df.sections.map(lambda x: x["text"], na_action = "ignore"))\
.drop(columns = "sections")\
.assign(heading = lambda df: df.heading.str.replace(r"\s+", " ", regex = True).str.strip(),
        text = lambda df: df.text.str.replace(r"\s+", " ", regex = True).str.strip())

show(paper_sections)

# writing to disk

final_citations.to_csv("../data/citations.csv", index = False)

final_papers.drop(columns = "sections").to_csv(
    "../data/papers.csv", 
    errors = "replace", 
    index = False
)

full_references.to_csv("../data/references.csv", index = False)

paper_sections.to_csv("../data/paper_sections.csv", index = False, errors = "replace")

# queries

unique_references.query("reference_id == 790")

references_abstract.query("reference_id == 2638")

full_references.query("reference_id == 790").abstract.tolist()

arxiv.query("title_clean == 'overviewofthetrec2019deeplearningtrack'")

arxiv.query("title_clean == 'overviewofthetrec2020deeplearningtrack'")
