==> notebooks/03_find_summary.ipynb <==
import malnis
from malnis import show
import pandas as pd
import importlib
from nltk.tokenize import sent_tokenize
from tqdm.auto import tqdm

data = pd.read_csv("../data/full_examples.csv", index_col = 0).head()
show(data)

importlib.reload(malnis)

for q, d, c in zip(tqdm(data["query"]), data.text, data.cited):
    s, m, d = malnis.find_summary(
        q, 
        d, 
        metric = "rouge-2", 
        component = "f", 
        starting_summary = c
    )
    print(q)
    print()
    print(s)
    print()
    print("max score:", m)
#     print(d)
    print("*" * 30)
    print()
