==> notebooks/04_data_split.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
# from sklearn.feature_extraction.text import TfidfVectorizer
from malnis import show
# from nltk.tokenize import word_tokenize
from tqdm.auto import tqdm
# import scipy.sparse as sp
# from sklearn.linear_model import LogisticRegression 
# from sklearn.ensemble import RandomForestClassifier
# from sklearn.svm import SVC
# from sklearn.decomposition import PCA, SparsePCA
# from sklearn.neural_network import MLPClassifier
# from sklearn.metrics import log_loss, PrecisionRecallDisplay, RocCurveDisplay
import numpy as np
# from sentence_transformers import SentenceTransformer
from sklearn.model_selection import train_test_split
# import torch
# from torch import nn
# import pytorch_lightning as pl
# import torch.utils.data as tud
# import seaborn as sns
# sns.set()

# The number of sentences in each paper goes from 59 to 4,447. I truncate to 512.

data = pd.read_pickle("../data/sentence_labels.pkl")\
.reset_index(drop = True)\
.assign(original_sentences = lambda df: df.sentences.map(len))\
.assign(
    sentences = lambda df: df.sentences.map(lambda y: y[:512]),
    relevance = lambda df: df.relevance.map(lambda y: y[:512]),    
)

show(data)

embeddings_folder = "/home/jarobyte/scratch/malnis_dataset/data/embeddings/specter/"

X = np.load(embeddings_folder + "X.npy")
X.shape

Y = np.load(embeddings_folder + "Y.npy")
Y.shape

(
    X_train, 
    X_devtest, 
    Y_train, 
    Y_devtest, 
    data_train, 
    data_devtest
) = train_test_split(
    X, 
    Y, 
    data, 
    random_state = 1,
    test_size = 0.2
)
print("X_train", X_train.shape)
print("X_devtest", X_devtest.shape)
print("Y_train", Y_train.shape)
print("Y_devtest", Y_devtest.shape)

(
    X_dev, 
    X_test, 
    Y_dev, 
    Y_test, 
    data_dev, 
    data_test
) = train_test_split(
    X_devtest, 
    Y_devtest, 
    data_devtest, 
    random_state = 1,
    test_size = 0.5
)
print("X_dev", X_dev.shape)
print("X_test", X_test.shape)
print("Y_dev", Y_dev.shape)
print("Y_test", Y_test.shape)

split_folder = "/home/jarobyte/scratch/malnis_dataset/data/"

to_write = [
    ("X_train", X_train),
    ("X_dev", X_dev),
    ("X_test", X_test),
    
    ("Y_train", Y_train),
    ("Y_dev", Y_dev),
    ("Y_test", Y_test),
]

for name, contents in to_write:
#     contents = contents.cpu().numpy()
    path = split_folder + name + ".npy"
    np.save(path, contents)
    print(name, "saved in", path)

to_write = [
    ("data_train", data_train),
    ("data_dev", data_dev),
    ("data_test", data_test),
]

for name, contents in to_write:
#     contents = contents.cpu().numpy()
    path = split_folder + name + ".pkl"
    contents.to_pickle(path)
    print(name, "saved in", path)
