==> notebooks/04_1_cosine_similarity.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
plt.style.use('seaborn-whitegrid')
from sklearn.feature_extraction.text import TfidfVectorizer
from tqdm.notebook import tqdm
from nltk.tokenize import sent_tokenize
from rouge import Rouge
from sklearn.model_selection import train_test_split
from sentence_transformers import SentenceTransformer
from sklearn.metrics import log_loss, PrecisionRecallDisplay, RocCurveDisplay
from malnis import show

train = pd.read_pickle("../data/split/labels_train.pkl")
show(train)

test = pd.read_pickle("../data/split/labels_test.pkl")
show(test)

true_targets = np.concatenate(test.relevance.to_list())
print("true targets", true_targets.shape)

records = test.to_dict("records")
len(records)

# TFIDF

# words

predictions = []
for d in tqdm(records):
    vectorizer = TfidfVectorizer()
    sentences = d["sentences"]
#     print(len(sentences))
    features = vectorizer.fit_transform(sentences)
#     print(features.shape)
    query_emb = vectorizer.transform([d["query"]])
#     print(query_emb.shape)
    relevance = (query_emb @ features.T).toarray().squeeze()
#     print(relevance.shape)
    predictions.append(relevance)

true_predictions = np.concatenate(predictions)
print("true predictions", true_predictions.shape)

np.save("evaluation/tfidf_words.npy", true_predictions)

# fig, axes = plt.subplots(1, 2, figsize = (12, 6))

# display = PrecisionRecallDisplay.from_predictions(
#     true_targets, true_predictions, name = "TFIDF Words",
#     ax = axes[0]
# )

# display = RocCurveDisplay.from_predictions(
#     true_targets, true_predictions, name = "TFIDF Words",
#     ax = axes[1]
# )

# chars

predictions = []
for d in tqdm(records):
    vectorizer = TfidfVectorizer(
        analyzer = "char",
        ngram_range = (3, 3)
    )    
    sentences = d["sentences"]
#     print(len(sentences))
    features = vectorizer.fit_transform(sentences)
#     print(features.shape)
    query_emb = vectorizer.transform([d["query"]])
#     print(query_emb.shape)
    relevance = (query_emb @ features.T).toarray().squeeze()
#     print(relevance.shape)
    predictions.append(relevance)

true_predictions = np.concatenate(predictions)
print("true predictions", true_predictions.shape)

np.save("evaluation/tfidf_chars.npy", true_predictions)

# fig, axes = plt.subplots(1, 2, figsize = (12, 6))

# display = PrecisionRecallDisplay.from_predictions(
#     true_targets, true_predictions, name = "TFIDF Chars",
#     ax = axes[0]
# )

# display = RocCurveDisplay.from_predictions(
#     true_targets, true_predictions, name = "TFIDF Chars",
#     ax = axes[1]
# )

# Sentence-BERT

model = SentenceTransformer(
    "sbert"
#     'all-MiniLM-L6-v2', 
#     cache_folder = "../assets"
#     "../cache/huggingface/transformers/"
#     cache_folder = "../cache/huggingface/transformers"
)
model

# sentence-bert

predictions = []
for d in tqdm(records):
#     vectorizer = TfidfVectorizer(
#         analyzer = "char",
#         ngram_range = (3, 3)
#     )    
    sentences = d["sentences"]
#     print(len(sentences))
#     features = vectorizer.fit_transform(sentences)
#     print(features.shape)
    features = model.encode(sentences)
    query_emb = model.encode([d["query"]])
#     query_emb = vectorizer.transform([d["query"]])
#     print(query_emb.shape)
    relevance = (query_emb @ features.T).squeeze()
#     print(relevance.shape)
    predictions.append(relevance)

true_predictions = np.concatenate(predictions)
print("true predictions", true_predictions.shape)

np.save("evaluation/sentence_bert.npy", true_predictions)

# fig, axes = plt.subplots(1, 2, figsize = (12, 6))

# display = PrecisionRecallDisplay.from_predictions(
#     true_targets, true_predictions, name = "Sentence-BERT",
#     ax = axes[0]
# )

# display = RocCurveDisplay.from_predictions(
#     true_targets, true_predictions, name = "Sentence-BERT",
#     ax = axes[1]
# )
