==> notebooks/05_evaluation.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import os
from sklearn.metrics import average_precision_score, roc_auc_score
from malnis import show

folder = "evaluation"

names = os.listdir(folder)
len(names)

predictions = [np.load(folder + "/" + m) for m in names]
len(predictions)

[x.shape for x in predictions]

[x.min() for x in predictions]

[x.max() for x in predictions]

test = pd.read_pickle("../data/split/labels_test.pkl")

targets = np.concatenate(test.relevance.to_list())
print("targets", targets.shape)

df = pd.DataFrame({"name":names, "predictions":predictions})\
.assign(
    name = lambda df: df.name.map(lambda x: x[:-4]),
    average_precision = lambda df: df.predictions.map(lambda x: average_precision_score(targets, x)),
    roc_auc = lambda df: df.predictions.map(lambda x: roc_auc_score(targets, x))
)\
# .drop(columns = "predictions")

show(df)
