==> notebooks/05_1_tfidf_mlp.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
plt.style.use('seaborn-whitegrid')
from sklearn.feature_extraction.text import TfidfVectorizer
from tqdm.notebook import tqdm
from sklearn.neural_network import MLPClassifier
# from nltk.tokenize import sent_tokenize
# from rouge import Rouge
# from sklearn.model_selection import train_test_split
# from sentence_transformers import SentenceTransformer
from sklearn.metrics import log_loss, roc_auc_score, average_precision_score
# PrecisionRecallDisplay, RocCurveDisplay
from malnis import show
import pickle

data_folder = "/home/jarobyte/scratch/malnis_dataset/data/"

train = pd.read_pickle(data_folder + "data_train.pkl")
show(train)

dev = pd.read_pickle(data_folder + "data_dev.pkl")
show(dev)

train_targets = np.concatenate(train.relevance.to_list())
train_targets.shape

output_folder = "/home/jarobyte/scratch/malnis_dataset/data/classical_models/"

# Words

%%time
# corpus = sum([d["sentences"] for d in tqdm(train_records)], start = [])
corpus = train.sentences.sum()
len(corpus)

%%time
vectorizer = TfidfVectorizer()
train_features = vectorizer.fit_transform(corpus)
train_features.shape

%%time 

n = 10000
X = train_features[:n]
Y = train_targets[:n]

clf = MLPClassifier(
    verbose = True, 
#     max_iter = 1
)
clf.partial_fit(X, Y, classes = np.unique(Y))

# predicting on dev set

dev_features = [vectorizer.transform(l) for l in tqdm(dev.sentences)]

dev_preds = [clf.predict_proba(f) for f in tqdm(dev_features)]

dev_targets = np.concatenate(dev.relevance.to_list())
dev_targets.shape

dev_preds_flat = np.concatenate(dev_preds)[:, 1]
dev_preds_flat.shape

average_precision_score(dev_targets, dev_preds_flat)

roc_auc_score(dev_targets, dev_preds_flat)

# saving to disk

with open("/home/jarobyte/scratch/malnis_dataset/words/predictions/test.pkl", "wb") as file:
    pickle.dump(clf, file)

with open("/home/jarobyte/scratch/malnis_dataset/words/predictions/test.pkl", "rb") as file:
    clf_2 = pickle.load(file)

dev_preds = [clf_2.predict_proba(f) for f in tqdm(dev_features)]

dev_preds_flat = np.concatenate(dev_preds)[:, 1]
dev_preds_flat.shape

average_precision_score(dev_targets, dev_preds_flat)

roc_auc_score(dev_targets, dev_preds_flat)
