==> notebooks/05_3_sentence_bert_concat_mlp.ipynb <==
import pandas as pd
import numpy as np
import torch
from torch import nn
import torch.utils.data as tud
import malnis
from random import randint
from sklearn.metrics import average_precision_score, roc_auc_score
from timeit import default_timer as timer
import importlib

data_folder = "/home/jarobyte/scratch/malnis_dataset/data/"

X_train = torch.tensor(np.load(data_folder + "embeddings/specter/X_train.npy"))#.cuda()
Y_train = torch.tensor(np.load(data_folder + "embeddings/specter/Y_train.npy"))#.cuda()

X_dev = torch.tensor(np.load(data_folder + "embeddings/specter/X_dev.npy"))
Y_dev = torch.tensor(np.load(data_folder + "embeddings/specter/Y_dev.npy"))

data_train = pd.read_pickle(data_folder + "data_train.pkl")
data_dev = pd.read_pickle(data_folder + "data_dev.pkl")

importlib.reload(malnis)
importlib.reload(malnis.models)

clf = malnis.models.MLP(input_size = 1536)

test = clf(X_train[:5])
test.shape, test.min(), test.max()

clf.fit(X_train[:100], Y_train[:100], epochs = 100)

test = clf(X_train[:5])
test.shape, test.min(), test.max()
