==> notebooks/09_no_augmentation_classical_models.ipynb <==
import pandas as pd
import numpy as np
from malnis import show
from nltk.tokenize import sent_tokenize
from tqdm.notebook import tqdm
import re
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics import average_precision_score, roc_auc_score

data = pd.read_pickle("../data/sentence_labels.pkl")\
.reset_index(drop = True)\
[["query", "sentences", "summary"]]\
.assign(summary = lambda df: df.summary.map(lambda x: sent_tokenize(re.sub(r"\.(?=[A-Z])", ". ", x[0]))))\

show(data)

df = data.sentences#.head(10)
clean_sentences = [[re.sub(r"\W", "", x.lower()) for x in l] for l in tqdm(df)]
clean_sentences[0][:10]

data.summary.map(len).value_counts()

df = data.summary#.head(10)
clean_summary = [[re.sub(r"\W", "", x.lower()) for x in l] for l in tqdm(df)]
clean_summary[0][:10]

pd.Series(clean_summary).map(len).value_counts()

data.sentences.map(len).sum()

relevance = [
    [True if s in summ else False for s in sents] 
    for sents, summ in zip(tqdm(clean_sentences), clean_summary)
]
sum([len(x) for x in relevance])

data = data.assign(relevance = relevance)
show(data)

all(data.relevance.map(len) == data.sentences.map(len))

data.relevance.map(any).value_counts()

data.relevance.map(any).sort_values()

filt = data.relevance.map(any)

new = data.loc[filt]
show(new)

train, test = train_test_split(new)
train.shape, test.shape

rel = np.concatenate(test.relevance.tolist())
rel.shape

%%time

corpus = train.sentences.sum()
len(corpus)

# words

vectorizer_words = TfidfVectorizer()
vectorizer_words.fit(corpus)

sentence_features_words = [vectorizer_words.transform(s) for s in tqdm(test.sentences)]
sentence_features_words[0].shape

# summary_features_words = [vectorizer_words.transform(s) for s in tqdm(test.summary)] 
# summary_features_words[0].shape

# scores_words = [
#     (sents @ summ.T).max(1).toarray().squeeze() 
#     for sents, summ in zip(tqdm(sentence_features_words), summary_features_words)
# ]
# scores_words[0].shape

query_features_words = [vectorizer_words.transform([s]) for s in tqdm(test["query"])]

len(query_features_words), query_features_words[0].shape 

scores_words = [
    (sents @ q.T).toarray().squeeze() 
    for sents, q in zip(tqdm(sentence_features_words), query_features_words)
]
len(scores_words), scores_words[0].shape

test.sentences.map(len).sum()

final_scores_words = np.concatenate(scores_words)
final_scores_words.shape

average_precision_score(rel, final_scores_words)

roc_auc_score(rel, final_scores_words)

# chars

vectorizer_chars = TfidfVectorizer(analyzer = "char", ngram_range = (3, 3))
vectorizer_chars.fit(corpus)

sentence_features_chars = [vectorizer_chars.transform(s) for s in tqdm(test.sentences)]
sentence_features_chars[0].shape

# summary_features_chars = [vectorizer_chars.transform(s) for s in tqdm(test.summary)] 
# summary_features_chars[0].shape

# scores_chars = [
#     (sents @ summ.T).max(1).toarray().squeeze() 
#     for sents, summ in zip(tqdm(sentence_features_chars), summary_features_chars)
# ]
# scores_chars[0].shape

query_features_chars = [vectorizer_chars.transform([s]) for s in tqdm(test["query"])] 
query_features_chars[0].shape

scores_chars = [
    (sents @ q.T).toarray().squeeze() 
    for sents, q in zip(tqdm(sentence_features_chars), query_features_chars)
]
scores_chars[0].shape

test.sentences.map(len).sum()

final_scores_chars = np.concatenate(scores_chars)
final_scores_chars.shape

average_precision_score(rel, final_scores_chars)

roc_auc_score(rel, final_scores_chars)

# queries

i = 10
list(zip(test.relevance.tolist()[i], scores_words[i].round(2)))

i = 30
list(zip(test.relevance.tolist()[i], scores_chars[i].round(2)))

any(relevance[30])
