==> notebooks/10_quotes_preprocessing.ipynb <==
import os
import json
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm
from malnis import show

files = [x for x in os.listdir("quotes") if "json" in x]
files

dicts = []
for f in tqdm(files):
    with open("quotes/" + f, "r") as file:
        dicts.append(json.loads(file.read()))

raw = pd.DataFrame(dicts, columns = list(dicts[0].keys()))
show(raw)

dfs = raw[["history", "query"]]\
.assign(history = lambda df: df.history.map(pd.read_json))\
.assign(
    history = lambda df: df.apply(
        lambda r: r["history"].assign(query = r["query"])\
        .sort_values(["filename", "sentence"]).groupby("filename").agg(list).reset_index(),
        axis = 1
    )
)

show(dfs)

data = pd.concat(dfs.history.tolist())\
.assign(
    query = lambda df: df["query"].map(min),
    relevance = lambda df: df.relevance.map(lambda x: [1 if y in (1, True) else 0 for y in x])
)\
[["text", "query", "relevance"]]\
.reset_index(drop = True)

show(data, 30)

sum([pd.read_json(d["history"]).shape[0] for d in dicts])

data.text.map(len).sum()

f = data.relevance.map(lambda x: all([0 if y == 1 else 1 for y in x]) or all(x))
f

data.loc[~f].to_pickle("quotes.pkl")
