==> notebooks/10_quotes_results_classical.ipynb <==
import pandas as pd
import numpy as np
from malnis import show
from sklearn.metrics import average_precision_score, roc_auc_score
from tqdm.notebook import tqdm
from sklearn.feature_extraction.text import TfidfVectorizer
import pickle

data = pd.read_pickle("quotes.pkl")
show(data, 30)

train = pd.read_pickle("/home/jarobyte/scratch/malnis_dataset/data/data_train.pkl")
show(train)

results = []

# classical models

%time
corpus = train.sentences.sum()
len(corpus)

## words

word_vectorizer = TfidfVectorizer(
    analyzer = "word",
    ngram_range = (1, 1)
)
word_vectorizer.fit(corpus)

word_features = data.text.map(word_vectorizer.transform)
len(word_features)

word_queries = data["query"].map(lambda x: word_vectorizer.transform([x]))
len(word_queries)

### tfidf

word_scores = [(f @ q.T).toarray().squeeze() for f, q in zip(word_features, word_queries)]
len(word_scores)

word_scores[0]

word_ap = [average_precision_score(l, s) for l, s in zip(tqdm(data.relevance), word_scores)]
len(word_ap)

word_ra = [roc_auc_score(l, s) for l, s in zip(tqdm(data.relevance), word_scores)]
len(word_ra)

results.append(
    dict(
        model = "tfidf_words",
        ap_mean = np.mean(word_ap), 
        ap_std = 2 * np.std(word_ap)/np.sqrt(len(word_ap)),
        ra_mean = np.mean(word_ra), 
        ra_std = 2 * np.std(word_ra)/np.sqrt(len(word_ra))
    )
)
results

## mlp

with open("/home/jarobyte/scratch/malnis_dataset/mlp_tfidf/words/models/15010241_7.pkl", "rb") as file:
    mlp_words = pickle.load(file)

word_mlp_scores = [mlp_words.predict_proba(x)[:, 1] for x in word_features]
len(word_mlp_scores)

word_mlp_ap = [average_precision_score(l, s) for l, s in zip(tqdm(data.relevance), word_mlp_scores)]
len(word_mlp_ap)

word_mlp_ra = [roc_auc_score(l, s) for l, s in zip(tqdm(data.relevance), word_mlp_scores)]
len(word_mlp_ra)

results.append(
    dict(
        model = "mlp_words",
        ap_mean = np.mean(word_mlp_ap), 
        ap_std = 2 * np.std(word_mlp_ap)/np.sqrt(len(word_mlp_ap)),
        ra_mean = np.mean(word_mlp_ra), 
        ra_std = 2 * np.std(word_mlp_ra)/np.sqrt(len(word_mlp_ra))
    )
)
results

## chars

char_vectorizer = TfidfVectorizer(
    analyzer = "char",
    ngram_range = (3, 3)
)
char_vectorizer.fit(corpus)

char_features = data.text.map(char_vectorizer.transform)
len(char_features)

char_queries = data["query"].map(lambda x: char_vectorizer.transform([x]))
len(char_queries)

### tfidf

char_scores = [(f @ q.T).toarray().squeeze() for f, q in zip(char_features, char_queries)]
len(char_scores)

char_scores[0]

char_ap = [average_precision_score(l, s) for l, s in zip(tqdm(data.relevance), char_scores)]
len(char_ap)

char_ra = [roc_auc_score(l, s) for l, s in zip(tqdm(data.relevance), char_scores)]
len(char_ra)

results.append(
    dict(
        model = "tfidf_chars",
        ap_mean = np.mean(char_ap), 
        ap_std = 2 * np.std(char_ap)/np.sqrt(len(char_ap)),
        ra_mean = np.mean(char_ra), 
        ra_std = 2 * np.std(char_ra)/np.sqrt(len(char_ra))
    )
)
results

## mlp

with open("/home/jarobyte/scratch/malnis_dataset/mlp_tfidf/chars/models/14986643_2.pkl", "rb") as file:
    mlp_chars = pickle.load(file)

char_mlp_scores = [mlp_chars.predict_proba(x)[:, 1] for x in char_features]
len(char_mlp_scores)

char_mlp_ap = [average_precision_score(l, s) for l, s in zip(tqdm(data.relevance), char_mlp_scores)]
len(char_mlp_ap)

char_mlp_ra = [roc_auc_score(l, s) for l, s in zip(tqdm(data.relevance), char_mlp_scores)]
len(char_mlp_ra)

results.append(
    dict(
        model = "mlp_chars",
        ap_mean = np.mean(char_mlp_ap), 
        ap_std = 2 * np.std(char_mlp_ap)/np.sqrt(len(char_mlp_ap)),
        ra_mean = np.mean(char_mlp_ra), 
        ra_std = 2 * np.std(char_mlp_ra)/np.sqrt(len(char_mlp_ra))
    )
)
results

results_df = pd.DataFrame(results)
show(results_df)

print(results_df.round(3).to_latex())

results_df.to_pickle("quotes_results_classical.pkl")
