==> notebooks/10_quotes_results_deep.ipynb <==
import pandas as pd
import numpy as np
from malnis import show
import malnis
from sklearn.metrics import average_precision_score, roc_auc_score
from sentence_transformers import SentenceTransformer
from tqdm.notebook import tqdm
import torch

data = pd.read_pickle("quotes.pkl")
show(data, 30)

results = []

# sentence-bert


model = SentenceTransformer(
    "sbert"
#     'all-MiniLM-L6-v2', 
#     cache_folder = "../assets"
#     "../cache/huggingface/transformers/"
#     cache_folder = "../cache/huggingface/transformers"
)

sbert_queries = [model.encode(t) for t in tqdm(data["query"])]

sbert_sents = [model.encode(t) for t in tqdm(data.text)]

sbert_features = [
    np.concatenate([np.tile(q, (l.shape[0], 1)), l], axis = 1)
    for q, l in zip(sbert_queries, sbert_sents)
]

len(sbert_features), sbert_features[0].shape

sbert_features = [torch.tensor(x).cuda() for x in sbert_features]

### cosine similarity

sbert_cosine_scores = [s @ q.T for s, q in zip(tqdm(sbert_sents), sbert_queries)]

sbert_cosine_ap = [average_precision_score(l, s) for l, s in zip(tqdm(data.relevance), sbert_cosine_scores)]
len(sbert_cosine_ap)

sbert_cosine_ra = [roc_auc_score(l, s) for l, s in zip(tqdm(data.relevance), sbert_cosine_scores)]
len(sbert_cosine_ra)

results.append(
    dict(
        model = "sbert_cosine",
        ap_mean = np.mean(sbert_cosine_ap), 
        ap_std = 2 * np.std(sbert_cosine_ap)/np.sqrt(len(sbert_cosine_ap)),
        ra_mean = np.mean(sbert_cosine_ra), 
        ra_std = 2 * np.std(sbert_cosine_ra)/np.sqrt(len(sbert_cosine_ra))
    )
)
results

### mlp

mlp_sbert = malnis.models.MLP(
    hidden_size = 300,
#     attention_heads = 4,
    num_layers = 3,
#     d_model = 1536
)
mlp_sbert.load_state_dict(
    torch.load("/home/jarobyte/scratch/malnis_dataset/mlp_pytorch/sbert/models/14628727_2.pt")
)
mlp_sbert.cuda()

mlp_sbert_scores = [mlp_sbert.predict(x) for x in sbert_features]
len(mlp_sbert_scores), mlp_sbert_scores[0].shape

mlp_sbert_ap = [average_precision_score(l, s) for l, s in zip(tqdm(data.relevance), mlp_sbert_scores)]
len(mlp_sbert_ap)

mlp_sbert_ra = [roc_auc_score(l, s) for l, s in zip(tqdm(data.relevance), mlp_sbert_scores)]
len(mlp_sbert_ra)

results.append(
    dict(
        model = "mlp_sbert",
        ap_mean = np.mean(mlp_sbert_ap), 
        ap_std = 2 * np.std(mlp_sbert_ap)/np.sqrt(len(mlp_sbert_ap)),
        ra_mean = np.mean(mlp_sbert_ra), 
        ra_std = 2 * np.std(mlp_sbert_ra)/np.sqrt(len(mlp_sbert_ra))
    )
)
results

### lstm

lstm_sbert = malnis.models.LSTM(
#     input_size = 1536,
    hidden_size = 500,
    num_layers = 3
)
lstm_sbert.load_state_dict(torch.load("/home/jarobyte/scratch/malnis_dataset/lstm/sbert/models/13932670_9.pt"))
lstm_sbert.cuda()

lstm_sbert_scores = [lstm_sbert.predict(x) for x in sbert_features]
len(lstm_sbert_scores), lstm_sbert_scores[0].shape

lstm_sbert_ap = [average_precision_score(l, s) for l, s in zip(tqdm(data.relevance), lstm_sbert_scores)]
len(lstm_sbert_ap)

lstm_sbert_ra = [roc_auc_score(l, s) for l, s in zip(tqdm(data.relevance), lstm_sbert_scores)]
len(lstm_sbert_ra)

results.append(
    dict(
        model = "lstm_sbert",
        ap_mean = np.mean(lstm_sbert_ap), 
        ap_std = 2 * np.std(lstm_sbert_ap)/np.sqrt(len(lstm_sbert_ap)),
        ra_mean = np.mean(lstm_sbert_ra), 
        ra_std = 2 * np.std(lstm_sbert_ra)/np.sqrt(len(lstm_sbert_ra))
    )
)
results

### transformer

transformer_sbert = malnis.models.Transformer(
    dim_feedforward = 250,
    attention_heads = 4,
    num_layers = 3
)
transformer_sbert.load_state_dict(
    torch.load("/home/jarobyte/scratch/malnis_dataset/transformer/sbert/models/13932680_10.pt")
)
transformer_sbert.cuda()

transformer_sbert_scores = [transformer_sbert.predict(x.unsqueeze(0)) for x in sbert_features]
len(transformer_sbert_scores), transformer_sbert_scores[0].shape

transformer_sbert_ap = [average_precision_score(l, s) for l, s in zip(tqdm(data.relevance), transformer_sbert_scores)]
len(transformer_sbert_ap)

transformer_sbert_ra = [roc_auc_score(l, s) for l, s in zip(tqdm(data.relevance), transformer_sbert_scores)]
len(transformer_sbert_ra)

results.append(
    dict(
        model = "transformer_sbert",
        ap_mean = np.mean(transformer_sbert_ap), 
        ap_std = 2 * np.std(transformer_sbert_ap)/np.sqrt(len(transformer_sbert_ap)),
        ra_mean = np.mean(transformer_sbert_ra), 
        ra_std = 2 * np.std(transformer_sbert_ra)/np.sqrt(len(transformer_sbert_ra))
    )
)
results

## specter

model = SentenceTransformer(
    "specter"
#     'all-MiniLM-L6-v2', 
#     cache_folder = "../assets"
#     "../cache/huggingface/transformers/"
#     cache_folder = "../cache/huggingface/transformers"
)

specter_queries = [model.encode(t) for t in tqdm(data["query"])]

specter_sents = [model.encode(t) for t in tqdm(data.text)]

specter_features = [
    np.concatenate([np.tile(q, (l.shape[0], 1)), l], axis = 1)
    for q, l in zip(specter_queries, specter_sents)
]

len(specter_features), specter_features[0].shape

specter_features = [torch.tensor(x).cuda() for x in specter_features]

### euclidean distance

specter_euclidean_scores = [2 / (1 + np.exp((s-q)**2).sum(1)) for s, q in zip(tqdm(specter_sents), specter_queries)]

specter_euclidean_ap = [average_precision_score(l, s) for l, s in zip(tqdm(data.relevance), specter_euclidean_scores)]
len(specter_euclidean_ap)

specter_euclidean_ra = [roc_auc_score(l, s) for l, s in zip(tqdm(data.relevance), specter_euclidean_scores)]
len(specter_euclidean_ra)

results.append(
    dict(
        model = "specter_euclidean",
        ap_mean = np.mean(specter_euclidean_ap), 
        ap_std = 2 * np.std(specter_euclidean_ap)/np.sqrt(len(specter_euclidean_ap)),
        ra_mean = np.mean(specter_euclidean_ra), 
        ra_std = 2 * np.std(specter_euclidean_ra)/np.sqrt(len(specter_euclidean_ra))
    )
)
results

### mlp

mlp_specter = malnis.models.MLP(
    hidden_size = 450,
#     attention_heads = 4,
    num_layers = 4,
    input_size = 1536
)
mlp_specter.load_state_dict(
    torch.load("/home/jarobyte/scratch/malnis_dataset/mlp_pytorch/specter/models/14678619_2.pt")
)
mlp_specter.cuda()

mlp_specter_scores = [mlp_specter.predict(x) for x in specter_features]
len(mlp_specter_scores), mlp_specter_scores[0].shape

mlp_specter_ap = [average_precision_score(l, s) for l, s in zip(tqdm(data.relevance), mlp_specter_scores)]
len(mlp_specter_ap)

mlp_specter_ra = [roc_auc_score(l, s) for l, s in zip(tqdm(data.relevance), mlp_specter_scores)]
len(mlp_specter_ra)

results.append(
    dict(
        model = "mlp_specter",
        ap_mean = np.mean(mlp_specter_ap), 
        ap_std = 2 * np.std(mlp_specter_ap)/np.sqrt(len(mlp_specter_ap)),
        ra_mean = np.mean(mlp_specter_ra), 
        ra_std = 2 * np.std(mlp_specter_ra)/np.sqrt(len(mlp_specter_ra))
    )
)
results

### lstm

lstm_specter = malnis.models.LSTM(
    input_size = 1536,
    hidden_size = 500,
    num_layers = 1
)
lstm_specter.load_state_dict(torch.load("/home/jarobyte/scratch/malnis_dataset/lstm/specter/models/13925873_9.pt"))
lstm_specter.cuda()

lstm_specter_scores = [lstm_specter.predict(x) for x in specter_features]
len(lstm_specter_scores), lstm_specter_scores[0].shape

lstm_specter_ap = [average_precision_score(l, s) for l, s in zip(tqdm(data.relevance), lstm_specter_scores)]
len(lstm_specter_ap)

lstm_specter_ra = [roc_auc_score(l, s) for l, s in zip(tqdm(data.relevance), lstm_specter_scores)]
len(lstm_specter_ra)

results.append(
    dict(
        model = "lstm_specter",
        ap_mean = np.mean(lstm_specter_ap), 
        ap_std = 2 * np.std(lstm_specter_ap)/np.sqrt(len(lstm_specter_ap)),
        ra_mean = np.mean(lstm_specter_ra), 
        ra_std = 2 * np.std(lstm_specter_ra)/np.sqrt(len(lstm_specter_ra))
    )
)
results

### transformer

transformer_specter = malnis.models.Transformer(
    dim_feedforward = 350,
    attention_heads = 4,
    num_layers = 3,
    d_model = 1536
)
transformer_specter.load_state_dict(
    torch.load("/home/jarobyte/scratch/malnis_dataset/transformer/specter/models/14158104_4.pt")
)
transformer_specter.cuda()

transformer_specter_scores = [transformer_specter.predict(x.unsqueeze(0)) for x in specter_features]
len(transformer_specter_scores), transformer_specter_scores[0].shape

transformer_specter_ap = [average_precision_score(l, s) for l, s in zip(tqdm(data.relevance), transformer_specter_scores)]
len(transformer_specter_ap)

transformer_specter_ra = [roc_auc_score(l, s) for l, s in zip(tqdm(data.relevance), transformer_specter_scores)]
len(transformer_specter_ra)

results.append(
    dict(
        model = "transformer_specter",
        ap_mean = np.mean(transformer_specter_ap), 
        ap_std = 2 * np.std(transformer_specter_ap)/np.sqrt(len(transformer_specter_ap)),
        ra_mean = np.mean(transformer_specter_ra), 
        ra_std = 2 * np.std(transformer_specter_ra)/np.sqrt(len(transformer_specter_ra))
    )
)
results

results_df = pd.DataFrame(results)
show(results_df)

print(results_df.round(3).to_latex())

results_df.to_pickle("quotes_results_deep.pkl")
