==> notebooks/11_quotes_results.ipynb <==
import pandas as pd
from malnis import show

classical = pd.read_pickle("quotes_results_classical.pkl")
show(classical)

deep = pd.read_pickle("quotes_results_deep.pkl")
show(deep)

full = pd.concat([classical, deep]).sort_values("ra_mean", ascending = False)
show(full, 10)

print(full.set_index("model").round(3).to_latex())
